Added: tab wise permission restriction on the hotel feature delete process

+++ modules/hotelreservationsystem/controllers/admin/AdminHotelfeaturesController.php
@@ -285,13 +285,20 @@ class AdminHotelFeaturesController extends ModuleAdminController
 
     public function ajaxProcessDeleteFeature()
     {
-        $idFeature = Tools::getValue('feature_id');
-        $objHotelFeatures = new HotelFeatures();
-        if ($objHotelFeatures->deleteHotelFeatures($idFeature)) {
-            die('success');
+        $response = array('status' => false);
+        if ($this->tabAccess['delete']) {
+            $idFeature = Tools::getValue('feature_id');
+            $objHotelFeatures = new HotelFeatures();
+            if ($objHotelFeatures->deleteHotelFeatures($idFeature)) {
+                $response['status'] = true;
+            } else {
+                $response['msg'] = $this->l('Some error occurred while deleting feature. Please try again.', null, true);
+            }
         } else {
-            echo 0;
+            $response['msg'] = $this->l('You do not have the permission to delete this.', null, true);
         }
+
+        $this->ajaxDie(json_encode($response));
     }
 
     public function setMedia()
